settings.js moved to assets/js
