+++ assets/js/settings.js
