+++ src/blocks/button/index.js
@@ -1,4 +1,5 @@
-import { button } from "../settings";
+import { button } from "../../../assets/js/settings";
+
 import edit from "./edit";
 import save from "./save";
 import deprecatedContent from "./deprecated";
