+++ src/blocks/dual-heading/index.js
@@ -1,4 +1,4 @@
-import { dualHeading } from "../settings";
+import { dualHeading } from "../../../assets/js/settings";
 import PbgIcon from "../icons";
 import edit from "./edit";
 import save from "./save";
