+++ src/blocks/icon-box/index.js
@@ -1,4 +1,4 @@
-import { iconBox } from "../settings";
+import { iconBox } from "../../../assets/js/settings";
 import PbgIcon from "../icons";
 import edit from "./edit";
 import save from "./save";
