+++ src/blocks/video-box/index.js
@@ -1,4 +1,4 @@
-import { videoBox } from "../settings";
+import { videoBox } from "../../../assets/js/settings";
 import PbgIcon from "../icons";
 import edit from "./edit";
 import save from "./save";
